requirements.txt
flask
requests
beautifulsoup4
fpdf2
gunicorn